Update 10x_Experimental_Scenarios.md
Adding the 10x lane name to the ilbrary and FASTQ names (e.g. L1-XL1-GEX, FQ1-XL1-GEX). This is because the 10x lane determines the set of cells that are measured. This allows us to run the same sample across multiple lanes (to get independent sets of cells) and keep track of this in the name