added HTO to HIMC Sample sheet
the HIMC sample sheet is the primary spreadsheet that is being handed from the 10x team to the computational team - I added the missing HTO data. The spreadsheet combines sample level and FASTQ level data as well as information that Cell Ranger is kept in the dark about (e.g. HTOs) but that need to be retained during data processing for later steps.